+++ exporters/ostream/test/ostream_log_test.cc
@@ -249,7 +249,9 @@ TEST(OStreamLogExporter, IntegrationTest)
   auto apiProvider = nostd::shared_ptr<logs_api::LoggerProvider>(sdkProvider);
   auto provider    = nostd::shared_ptr<logs_api::LoggerProvider>(apiProvider);
   logs_api::Provider::SetLoggerProvider(provider);
-  auto logger = logs_api::Provider::GetLoggerProvider()->GetLogger("Logger");
+  const std::string schema_url{"https://opentelemetry.io/schemas/1.2.0"};
+  auto logger =
+      logs_api::Provider::GetLoggerProvider()->GetLogger("Logger", "", "lib_name", "", schema_url);
 
   // Back up cout's streambuf
   std::streambuf *original = std::cout.rdbuf();